pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "cx-agent-studio-cicd"
version = "1.0.0"
description = "CI/CD pipeline for Google Cloud CX Agent Studio (Gemini Enterprise for Customer Experience)"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "google-cloud-storage>=2.14.0",
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-core>=2.17.0",
    "google-cloud-resource-manager>=1.12.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.12",
    "types-requests>=2.31.0",
]

[project.scripts]
cx-agent = "src.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "SIM", "TCH"]
ignore = ["S101", "S603", "S607", "TCH003"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
